move shared universal types into new file
